Scrolling Object Collection example scene polish.
