+++ Assets/MixedRealityToolkit.Examples/Experimental/ScrollingObjectCollection/Prefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 25e055ceadd0a6b45bf35abc99d7b94c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
