+++ Assets/MixedRealityToolkit.Examples/Experimental/ScrollingObjectCollection/Prefabs/ScrollingOC_Objects_Plated32x32.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: da1d75121e6bdea418a32232572019e0
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
