+++ Assets/MixedRealityToolkit.Examples/Experimental/ScrollingObjectCollection/Prefabs/ScrollingOC_PressableBtn_32x96.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f8e43a3ccc58ff943b0bed33339f073b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
