+++ Assets/MixedRealityToolkit.Examples/Experimental/ScrollingObjectCollection/Scenes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 98e055adf511eea4b877e074a0f783fe
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
