+++ Assets/MixedRealityToolkit.Examples/Experimental/ScrollingObjectCollection/Scenes/ScrollingObjectCollection.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 79a0bb4c0426b7a4099c35eeced52ee4
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
