+++ Assets/MixedRealityToolkit.Examples/Experimental/ScrollingObjectCollection/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c9935a531e49b2d46af0ebbeb485a1a0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
