+++ Assets/MixedRealityToolkit.Examples/Experimental/ScrollingObjectCollection/Scripts/ScrollableListPopulator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: da54a0d31c17eec4fad198a90dffab5e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
