--- Assets/MixedRealityToolkit.Examples/Experimental/ScrollingObjectCollection/ScrollingObjectCollection.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 79a0bb4c0426b7a4099c35eeced52ee4
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
