+++ Assets/MixedRealityToolkit.Examples/StandardAssets/Materials/ExampleBoxWithBorder.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6438f92337e03964fb6094158cb3a8b4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
